moved linreg schema
