+++ src/main/scala/is/hail/rest/RestService.scala
@@ -14,10 +14,139 @@ import org.json4s.NoTypeHints
 import org.json4s.jackson.Serialization
 import org.json4s.jackson.Serialization.write
 
+import scala.collection.mutable
 import scala.concurrent.ExecutionContext
 
 object RestService {
-  def inRangeFunc(n: Int, minMAC: Int, maxMAC: Int): Double => Boolean = {
+  // construct (covNames, covVariants)
+  def getCovariates(
+    phenotypeOpt: Option[String],
+    covariatesOpt: Option[Array[Covariate]],
+    availableCovariates: Set[String]): (Array[String], Array[Variant]) = {
+
+    val covNamesSet = mutable.Set[String]()
+    val covVariantsSet = mutable.Set[Variant]()
+    
+    covariatesOpt.foreach { covariates =>
+      for (c <- covariates)
+        c.`type` match {
+          case "phenotype" =>
+            c.name match {
+              case Some(name) =>
+                if (availableCovariates(name))
+                  if (covNamesSet(name))
+                    throw new RestFailure(s"Covariate $name is included as a covariate more than once")
+                  else
+                    covNamesSet += name
+                else
+                  throw new RestFailure(s"$name is not a valid covariate name")
+              case None =>
+                throw new RestFailure("Covariate of type 'phenotype' must include 'name' field in request")
+            }
+          case "variant" =>
+            (c.chrom, c.pos, c.ref, c.alt) match {
+              case (Some(chr), Some(pos), Some(ref), Some(alt)) =>
+                val v = Variant(chr, pos, ref, alt)
+                if (covVariantsSet(v))
+                  throw new RestFailure(s"$v is included as a covariate more than once")
+                else
+                  covVariantsSet += v
+              case missingFields =>
+                throw new RestFailure("Covariate of type 'variant' must include 'chrom', 'pos', 'ref', and 'alt' fields in request")
+            }
+          case other =>
+            throw new RestFailure(s"Covariate type must be 'phenotype' or 'variant': got $other")
+        }
+    }
+        
+    phenotypeOpt.foreach { phenotype =>
+      if (covNamesSet(phenotype))
+        throw new RestFailure(s"$phenotype appears as both response phenotype and covariate")
+      if (!availableCovariates(phenotype))
+        throw new RestFailure(s"$phenotype is not a valid phenotype name")
+    }
+    
+    (covNamesSet.toArray, covVariantsSet.toArray)
+  }
+  
+  // construct (window, chrom, minPos, maxPos, minMAC, maxMAC)
+  def getWindow(variant_filters: Option[Array[VariantFilter]], maxWidth: Int): (Interval[Locus], String, Int, Int, Int, Int) = {
+        // construct variant filters
+    var chrom = ""
+
+    var minPos = 1
+    var maxPos = Int.MaxValue // 2,147,483,647 is greater than length of longest chromosome
+
+    var minMAC = 1
+    var maxMAC = Int.MaxValue
+
+    val nonNegIntRegEx = """\d+""".r
+
+    variant_filters.foreach(_.foreach { f =>
+      f.operand match {
+        case "chrom" =>
+          if (!(f.operator == "eq" && f.operand_type == "string"))
+            throw new RestFailure(s"chrom filter operator must be 'eq' and operand_type must be 'string': got '${ f.operator }' and '${ f.operand_type }'")
+          else if (f.value.isEmpty)
+            throw new RestFailure("chrom filter value cannot be the empty string")
+          else if (chrom.isEmpty)
+            chrom = f.value
+          else if (chrom != f.value)
+            throw new RestFailure(s"Got incompatible chrom filters: '$chrom' and '${ f.value }'")
+        case "pos" =>
+          if (f.operand_type != "integer")
+            throw new RestFailure(s"pos filter operand_type must be 'integer': got '${ f.operand_type }'")
+          else if (!nonNegIntRegEx.matches(f.value))
+            throw new RestFailure(s"Value of position in variant_filter must be a valid non-negative integer: got '${ f.value }'")
+          else {
+            val pos = f.value.toInt
+            f.operator match {
+              case "gte" => minPos = minPos max pos
+              case "gt" => minPos = minPos max (pos + 1)
+              case "lte" => maxPos = maxPos min pos
+              case "lt" => maxPos = maxPos min (pos - 1)
+              case "eq" => minPos = minPos max pos; maxPos = maxPos min pos
+              case other =>
+                throw new RestFailure(s"pos filter operator must be 'gte', 'gt', 'lte', 'lt', or 'eq': got '$other'")
+            }
+          }
+        case "mac" =>
+          if (f.operand_type != "integer")
+            throw new RestFailure(s"mac filter operand_type must be 'integer': got '${ f.operand_type }'")
+          else if (!nonNegIntRegEx.matches(f.value))
+            throw new RestFailure(s"mac filter value must be a valid non-negative integer: got '${ f.value }'")
+          val mac = f.value.toInt
+          f.operator match {
+            case "gte" => minMAC = minMAC max mac
+            case "gt" => minMAC = minMAC max (mac + 1)
+            case "lte" => maxMAC = maxMAC min mac
+            case "lt" => maxMAC = maxMAC min (mac - 1)
+            case other =>
+              throw new RestFailure(s"mac filter operator must be 'gte', 'gt', 'lte', 'lt': got '$other'")
+          }
+        case other => throw new RestFailure(s"Filter operand must be 'chrom' or 'pos': got '$other'")
+      }
+    })
+
+    // construct window
+    if (chrom.isEmpty)
+      throw new RestFailure("No chromosome specified in variant_filter")
+
+    val width = maxPos - minPos
+    if (width > maxWidth)
+      throw new RestFailure(s"Interval length cannot exceed $maxWidth: got $width")
+
+    if (width < 0)
+      throw new RestFailure(s"Window is empty: got start $minPos and end $maxPos")
+
+    val window = Interval(Locus(chrom, minPos), Locus(chrom, maxPos + 1))
+
+    info(s"Using window ${ RestService.windowToString(window) } of size ${ width + 1 }")
+    
+    (window, chrom, minPos, maxPos, minMAC, maxMAC)
+  }
+  
+  def makeFilterAC(n: Int, minMAC: Int, maxMAC: Int): Double => Boolean = {
     val lowerMinAC = 1 max minMAC
     val lowerMaxAC = n min maxMAC
     val upperMinAC = 2 * n - lowerMaxAC
